+++ udesksdk/UdeskSDKUI/src/main/java/cn/udesk/aac/UdeskViewMode.java
@@ -137,7 +137,7 @@ public class UdeskViewMode extends ViewModel {
         sendMessageLiveData.setCustomerId(customerId);
         robotApiData.setCustomerId(customerId);
     }
-    public void setSessionId(int sessionId){
+    public void setSessionId(long sessionId) {
         robotApiData.setSessionId(sessionId);
     }
